Remove meaningless numerical instances from PPosixTime
