+++ Plutarch/Enum.hs
@@ -9,9 +9,7 @@ import Plutarch.Builtin.Bool (pif)
 import Plutarch.Builtin.Integer (PInteger)
 import Plutarch.Internal.Eq ((#==))
 import Plutarch.Internal.Fix (pfix)
-import Plutarch.Internal.Numeric ()
-import Plutarch.Internal.Numeric.Additive (PPositive, (#+))
-import Plutarch.Internal.Numeric.Multiplicative (pone)
+import Plutarch.Internal.Numeric (PPositive, pone, (#+))
 import Plutarch.Internal.Ord (POrd)
 import Plutarch.Internal.Other (pto)
 import Plutarch.Internal.PLam (plam)
