+++ Plutarch/Internal/TryFrom.hs
@@ -15,6 +15,8 @@ module Plutarch.Internal.TryFrom (
   pdowncastF,
 ) where
 
+import Data.Functor.Const (Const)
+import GHC.Generics (Generic)
 import Plutarch.Builtin.Bool (PBool, pif, (#||))
 import Plutarch.Builtin.ByteString (PByteString)
 import Plutarch.Builtin.Data (
@@ -36,12 +38,6 @@ import Plutarch.Builtin.Integer (
   peqInteger,
  )
 import Plutarch.Builtin.String (ptraceInfo)
-import Plutarch.Internal.Numeric.Additive (PPositive, ptryPositive)
-
--- import Plutarch.Builtin.Unit
-
-import Data.Functor.Const (Const)
-import GHC.Generics (Generic)
 import Plutarch.Internal.IsData (
   PIsData,
   pdata,
@@ -49,6 +45,7 @@ import Plutarch.Internal.IsData (
   pfromData,
  )
 import Plutarch.Internal.ListLike (PListLike (pnull), pmap)
+import Plutarch.Internal.Numeric (PPositive, ptryPositive)
 import Plutarch.Internal.PLam (PLamN (plam))
 import Plutarch.Internal.PlutusType (PInner)
 import Plutarch.Internal.Subtype (
