+++ plutarch-ledger-api/src/Plutarch/LedgerApi/Utils.hs
@@ -45,7 +45,7 @@ import Data.Bifunctor (first)
 import Data.Kind (Type)
 import GHC.Generics (Generic)
 import GHC.Records (getField)
-import Plutarch.Internal.Numeric.Additive (PPositive)
+import Plutarch.Internal.Numeric (PPositive)
 import Plutarch.Internal.PlutusType (PlutusType (pcon', pmatch'))
 import Plutarch.Prelude
 import Plutarch.Unsafe (punsafeCoerce)
