+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Time.hs
@@ -40,7 +40,7 @@ newtype PPosixTime (s :: S) = PPosixTime (Term s (PDataNewtype PInteger))
 -- | @since WIP
 instance PCountable PPosixTime where
   {-# INLINEABLE psuccessor #-}
-  psuccessor = phoistAcyclic $ plam (#+ pone)
+  psuccessor = phoistAcyclic $ plam (\x -> x #+ pposixTime pone)
   {-# INLINEABLE psuccessorN #-}
   psuccessorN = phoistAcyclic $ plam $ \p t ->
     let p' = pcon . PPosixTime . pcon . PDataNewtype . pdata . pto $ p
@@ -49,26 +49,11 @@ instance PCountable PPosixTime where
 -- | @since WIP
 instance PEnumerable PPosixTime where
   {-# INLINEABLE ppredecessor #-}
-  ppredecessor = phoistAcyclic $ plam (#- pone)
+  ppredecessor = phoistAcyclic $ plam (\x -> x #- pposixTime pone)
   {-# INLINEABLE ppredecessorN #-}
   ppredecessorN = phoistAcyclic $ plam $ \p t ->
     let p' = pcon . PPosixTime . pcon . PDataNewtype . pdata . pto $ p
      in t #- p'
-
--- | @since 2.0.0
-instance PIntegral PPosixTime where
-  {-# INLINEABLE pdiv #-}
-  pdiv = phoistAcyclic $ plam $ \t1 t2 ->
-    pposixTime (pdiv # unPPosixTime t1 # unPPosixTime t2)
-  {-# INLINEABLE pmod #-}
-  pmod = phoistAcyclic $ plam $ \t1 t2 ->
-    pposixTime (pmod # unPPosixTime t1 # unPPosixTime t2)
-  {-# INLINEABLE pquot #-}
-  pquot = phoistAcyclic $ plam $ \t1 t2 ->
-    pposixTime (pquot # unPPosixTime t1 # unPPosixTime t2)
-  {-# INLINEABLE prem #-}
-  prem = phoistAcyclic $ plam $ \t1 t2 ->
-    pposixTime (prem # unPPosixTime t1 # unPPosixTime t2)
 
 -- | @since WIP
 instance PAdditiveSemigroup PPosixTime where
@@ -89,28 +74,6 @@ instance PAdditiveGroup PPosixTime where
   pnegate = phoistAcyclic $ plam $ \t -> pposixTime (pnegate # unPPosixTime t)
   {-# INLINEABLE (#-) #-}
   t1 #- t2 = pposixTime (unPPosixTime t1 #- unPPosixTime t2)
-
--- | @since WIP
-instance PMultiplicativeSemigroup PPosixTime where
-  {-# INLINEABLE (#*) #-}
-  t1 #* t2 = pposixTime (unPPosixTime t1 #* unPPosixTime t2)
-
--- | @since WIP
-instance PMultiplicativeMonoid PPosixTime where
-  {-# INLINEABLE pone #-}
-  pone = pposixTime pone
-
--- | @since WIP
-instance PIntegralDomain PPosixTime where
-  {-# INLINEABLE psignum #-}
-  psignum = phoistAcyclic $ plam $ \t -> pposixTime (psignum # unPPosixTime t)
-  {-# INLINEABLE pabs #-}
-  pabs = phoistAcyclic $ plam $ \t -> pposixTime (pabs # unPPosixTime t)
-
--- | @since WIP
-instance PRing PPosixTime where
-  {-# INLINEABLE pfromInteger #-}
-  pfromInteger = pposixTime . pconstant
 
 -- | @since 2.0.0
 instance DerivePlutusType PPosixTime where
