+++ plutarch-testlib/goldens/show.bench.golden
@@ -1,7 +1,7 @@
 unit {"exBudgetCPU":16100,"exBudgetMemory":200,"scriptSizeBytes":11}
 bool.true {"exBudgetCPU":284149,"exBudgetMemory":1401,"scriptSizeBytes":33}
 bool.false {"exBudgetCPU":284149,"exBudgetMemory":1401,"scriptSizeBytes":33}
-int.0 {"exBudgetCPU":3230792,"exBudgetMemory":14417,"scriptSizeBytes":304}
+int.0 {"exBudgetCPU":3657208,"exBudgetMemory":15821,"scriptSizeBytes":304}
 int.5 {"exBudgetCPU":4832702,"exBudgetMemory":20427,"scriptSizeBytes":304}
 int.-5 {"exBudgetCPU":5319075,"exBudgetMemory":21832,"scriptSizeBytes":304}
 int.10 {"exBudgetCPU":6370737,"exBudgetMemory":26243,"scriptSizeBytes":304}
